Split off ddate module
