+++ src/altdate.rs
@@ -1,0 +1,12 @@
+
+pub mod ddate;
+
+pub mod altdate {
+
+#[allow(dead_code)]
+#[derive(Debug)]
+    enum Calendar {
+        Discordian,
+    }
+}
+
